Spitting marker database algos
