+++ crates/paralegal-flow/src/ann/db/mod.rs
@@ -13,39 +13,32 @@
 use crate::{
     ann::{Annotation, MarkerAnnotation},
     args::{Args, Stub},
-    utils::{
-        func_of_term, resolve::expect_resolve_string_to_def_id, type_for_constructor, FunctionKind,
-        InstanceExt, IntoDefId, TyExt,
-    },
-    Either, HashMap, HashSet,
+    utils::{resolve::expect_resolve_string_to_def_id, FunctionKind, InstanceExt, IntoDefId},
+    Either, HashMap,
 };
-use anyhow::Context;
-use flowistry::mir::FlowistryInput;
-use flowistry_pdg_construction::{
-    determine_async,
-    source_access::ParalegalDecoder,
-    source_access::{local_or_remote_paths, BodyCache},
-    utils::{handle_shims, is_virtual, try_monomorphize, try_resolve_function, ShimResult},
+use flowistry_pdg_construction::source_access::{
+    local_or_remote_paths, BodyCache, ParalegalDecoder,
 };
 use paralegal_spdg::Identifier;
 
-use rustc_data_structures::fx::{FxHashMap, FxHashSet};
+use rustc_data_structures::fx::FxHashMap;
 use rustc_errors::DiagMessage;
 use rustc_hir::def_id::{DefId, LOCAL_CRATE};
 use rustc_hir::{def::DefKind, def_id::CrateNum};
 use rustc_middle::{
-    mir::{self, Local},
-    ty::{self, GenericArgsRef, Instance, Ty, TyCtxt, TypingEnv},
+    mir::Local,
+    ty::{self, GenericArgsRef, Instance, Ty, TyCtxt},
 };
 use rustc_serialize::Decodable;
 
 use rustc_span::Span;
 use rustc_utils::cache::Cache;
+
+use anyhow::Context;
 use serde::Deserialize;
 use serde_json::json;
 
 use std::{
-    borrow::Cow,
     cell::{RefCell, RefMut},
     fs::File,
     io::Read,
@@ -53,6 +46,9 @@ use std::{
 };
 
 use super::{MarkerMeta, MarkerRefinement, MARKER_META_EXT};
+
+mod reachable;
+mod type_markers;
 
 type ExternalMarkers = HashMap<DefId, Vec<MarkerAnnotation>>;
 
@@ -77,12 +73,12 @@ impl<'tcx> From<MarkerDatabase<'tcx>> for MarkerCtx<'tcx> {
 
 impl<'tcx> MarkerCtx<'tcx> {
     #[inline]
-    fn tcx(&self) -> TyCtxt<'tcx> {
+    pub(super) fn tcx(&self) -> TyCtxt<'tcx> {
         self.0.tcx
     }
 
     #[inline]
-    fn db(&self) -> &MarkerDatabase<'tcx> {
+    pub(super) fn db(&self) -> &MarkerDatabase<'tcx> {
         &self.0
     }
 
@@ -182,62 +178,7 @@ impl<'tcx> MarkerCtx<'tcx> {
         self.is_marked(res.def_id()) || self.has_transitive_reachable_markers(res)
     }
 
-    /// Queries the transitive marker cache.
-    pub fn has_transitive_reachable_markers(
-        &self,
-        res: impl Into<MaybeMonomorphized<'tcx>>,
-    ) -> bool {
-        !self.get_reachable_markers(res).is_empty()
-    }
-
-    pub fn get_reachable_markers(&self, res: impl Into<MaybeMonomorphized<'tcx>>) -> &[Identifier] {
-        let res = res.into();
-        let def_id = res.def_id();
-        if self.is_marked(def_id) {
-            trace!("  Is marked");
-            return &[];
-        }
-        if is_virtual(self.tcx(), def_id) {
-            trace!("  Is virtual");
-            return &[];
-        }
-        if self.tcx().is_foreign_item(def_id) {
-            trace!("  Is foreign");
-            return &[];
-        }
-        if !(self.0.included_crates)(def_id.krate) {
-            trace!("  Is excluded");
-            return &[];
-        }
-        self.db()
-            .reachable_markers
-            .get_maybe_recursive(&res, |_| self.compute_reachable_markers(res))
-            .map_or(&[], Box::as_ref)
-    }
-
-    fn get_reachable_and_self_markers(
-        &self,
-        res: impl Into<MaybeMonomorphized<'tcx>>,
-    ) -> impl Iterator<Item = Identifier> + '_ {
-        let res = res.into();
-        let mut direct_markers = self
-            .combined_markers(res.def_id())
-            .map(|m| m.marker)
-            .peekable();
-        let is_self_marked = direct_markers.peek().is_some();
-        if is_self_marked {
-            let mut stat = self.borrow_function_marker_stat(res);
-            if stat.markers_on_self.is_empty() {
-                stat.markers_on_self
-                    .extend(self.combined_markers(res.def_id()).map(|m| m.marker));
-            }
-        }
-        let non_direct = (!is_self_marked).then(|| self.get_reachable_markers(res));
-
-        direct_markers.chain(non_direct.into_iter().flatten().copied())
-    }
-
-    fn borrow_function_marker_stat(
+    pub(super) fn borrow_function_marker_stat(
         &self,
         key: MaybeMonomorphized<'tcx>,
     ) -> RefMut<'_, FunctionMarkerStat<'tcx>> {
@@ -254,78 +195,7 @@ impl<'tcx> MarkerCtx<'tcx> {
         })
     }
 
-    /// If the transitive marker cache did not contain the answer, this is what
-    /// computes it.
-    fn compute_reachable_markers(&self, res: MaybeMonomorphized<'tcx>) -> Box<[Identifier]> {
-        trace!("Computing reachable markers for {res:?}");
-
-        if self.tcx().is_constructor(res.def_id()) {
-            let parent = type_for_constructor(self.tcx(), res.def_id());
-            return self
-                .combined_markers(parent)
-                .map(|m| m.marker)
-                .collect::<Box<_>>();
-        }
-        let body = self.0.body_cache.get(res.def_id());
-        let param_env = TypingEnv::post_analysis(self.tcx(), res.def_id())
-            .with_post_analysis_normalized(self.tcx());
-        let mono_body = match res {
-            MaybeMonomorphized::Monomorphized(res) => Cow::Owned(
-                try_monomorphize(
-                    res,
-                    self.tcx(),
-                    param_env,
-                    body.body(),
-                    self.tcx().def_span(res.def_id()),
-                )
-                .unwrap(),
-            ),
-            MaybeMonomorphized::Plain(_) => Cow::Borrowed(body.body()),
-        };
-        if let Some((async_fn, _, _)) = determine_async(self.tcx(), res.def_id(), &mono_body) {
-            self.borrow_function_marker_stat(res).is_async = Some(async_fn);
-            return self.get_reachable_markers(async_fn).into();
-        }
-        let expect_resolve = res.is_monomorphized();
-        let variable_markers = {
-            let mut stat = self.borrow_function_marker_stat(res);
-            mono_body
-                .local_decls
-                .iter_enumerated()
-                .flat_map(|(l, v)| {
-                    let markers = self.deep_type_markers(v.ty);
-                    if !markers.is_empty() {
-                        stat.markers_from_variables.push((
-                            l,
-                            v.ty,
-                            markers.iter().map(|v| v.1).collect(),
-                        ));
-                    }
-                    markers
-                })
-                .map(|m| m.1)
-                .collect::<Vec<_>>()
-        };
-
-        mono_body
-            .basic_blocks
-            .iter()
-            .flat_map(|bbdat| {
-                self.terminator_reachable_markers(
-                    res,
-                    &mono_body.local_decls,
-                    bbdat.terminator(),
-                    param_env,
-                    expect_resolve,
-                )
-            })
-            .chain(variable_markers)
-            .collect::<HashSet<_>>()
-            .into_iter()
-            .collect()
-    }
-
-    fn span_err(&self, span: Span, msg: impl Into<DiagMessage>) {
+    pub(super) fn span_err(&self, span: Span, msg: impl Into<DiagMessage>) {
         if self.0.config.relaxed() {
             self.tcx().dcx().span_warn(span, msg.into());
         } else {
@@ -340,224 +210,6 @@ impl<'tcx> MarkerCtx<'tcx> {
         } else {
             self.tcx().dcx().err(msg.into());
         }
-    }
-
-    /// Does this terminator carry a marker?
-    fn terminator_reachable_markers(
-        &self,
-        parent: MaybeMonomorphized<'tcx>,
-        local_decls: &mir::LocalDecls,
-        terminator: &mir::Terminator<'tcx>,
-        param_env: ty::TypingEnv<'tcx>,
-        expect_resolve: bool,
-    ) -> impl Iterator<Item = Identifier> + '_ {
-        let mut v = vec![];
-        trace!(
-            "  Finding reachable markers for terminator {:?}",
-            terminator.kind
-        );
-        let Some((def_id, gargs)) = func_of_term(self.tcx(), terminator) else {
-            return v.into_iter();
-        };
-        let mut res = if expect_resolve {
-            let Some(instance) = try_resolve_function(self.tcx(), def_id, param_env, gargs) else {
-                self.span_err(
-                        terminator.source_info.span,
-                        format!("cannot determine reachable markers, failed to resolve {def_id:?} with {gargs:?}")
-                    );
-                return v.into_iter();
-            };
-            let new_instance = match handle_shims(
-                instance,
-                self.tcx(),
-                param_env,
-                terminator.source_info.span,
-            ) {
-                ShimResult::IsHandledShim { instance, .. } => instance,
-                ShimResult::IsNonHandleableShim => {
-                    self.span_err(
-                        terminator.source_info.span,
-                        format!("cannot determine reachable markers, failed to handle shim {def_id:?} with {gargs:?}")
-                    );
-                    return v.into_iter();
-                }
-                ShimResult::IsNotShim => instance,
-            };
-            MaybeMonomorphized::Monomorphized(new_instance)
-        } else {
-            MaybeMonomorphized::Plain(def_id)
-        };
-        trace!(
-            "    Checking function {} for markers",
-            self.tcx().def_path_debug_str(res.def_id())
-        );
-
-        if let Some(model) = self.has_stub(res.def_id()) {
-            if let MaybeMonomorphized::Monomorphized(instance) = &mut res {
-                if let Ok(new_instance) = model.resolve_alternate_instance(
-                    self.tcx(),
-                    *instance,
-                    param_env,
-                    terminator.source_info.span,
-                ) {
-                    self.borrow_function_marker_stat(res).is_stubbed = Some(new_instance);
-                    v.extend(self.get_reachable_and_self_markers(new_instance));
-                }
-            } else {
-                self.span_err(
-                    terminator.source_info.span,
-                    "Could not apply stub to an partially resolved function",
-                );
-            };
-            return v.into_iter();
-        }
-
-        v.extend(self.get_reachable_and_self_markers(res));
-
-        // We have to proceed differently than graph construction,
-        // because we are not given the closure function, instead
-        // we are provided the id of the function that creates the
-        // future. As such we can't just monomorphize and traverse,
-        // we have to find the generator first.
-        if let ty::TyKind::Alias(ty::AliasTyKind::Opaque, alias) =
-            local_decls[mir::RETURN_PLACE].ty.kind()
-            && let ty::TyKind::Coroutine(closure_fn, substs) =
-                self.tcx().type_of(alias.def_id).skip_binder().kind()
-        {
-            trace!("    fits opaque type");
-            let async_closure_fn =
-                try_resolve_function(self.tcx(), *closure_fn, param_env, substs).unwrap();
-            v.extend(self.get_reachable_and_self_markers(async_closure_fn));
-            self.borrow_function_marker_stat(res).is_async = Some(async_closure_fn);
-        };
-        if !v.is_empty() {
-            self.borrow_function_marker_stat(parent)
-                .calls_with_reachable_markers
-                .push((res, terminator.source_info.span));
-        }
-        v.into_iter()
-    }
-
-    /// All the markers applied to this type and its subtypes.
-    ///
-    /// Returns `(ann, (ty, did))` tuples which are the marker annotation `ann`,
-    /// the specific type `ty` that it was applied to and the `did` [`Defid`] of
-    /// that type that was used to look up the annotations.
-    pub fn all_type_markers<'a>(
-        &'a self,
-        ty: ty::Ty<'tcx>,
-    ) -> impl Iterator<Item = (&'a MarkerAnnotation, (ty::Ty<'tcx>, DefId))> {
-        ty.walk().filter_map(|g| g.as_type()).flat_map(move |typ| {
-            typ.defid().into_iter().flat_map(move |did| {
-                self.combined_markers(did)
-                    .zip(std::iter::repeat((typ, did)))
-            })
-        })
-    }
-
-    pub fn shallow_type_markers<'a>(
-        &'a self,
-        key: ty::Ty<'tcx>,
-    ) -> impl Iterator<Item = TypeMarkerElem> + 'a {
-        use ty::*;
-        let def_id = match key.kind() {
-            Adt(def, _) => Some(def.did()),
-            Alias(_, inner) => Some(inner.def_id),
-            _ => None,
-        };
-        def_id
-            .map(|def_id| {
-                self.combined_markers(def_id)
-                    .map(move |m| (def_id, m.marker))
-            })
-            .into_iter()
-            .flatten()
-    }
-
-    pub fn deep_type_markers<'a>(&'a self, key: ty::Ty<'tcx>) -> &'a TypeMarkers {
-        self.0
-            .type_markers
-            .get_maybe_recursive(&key, |key| {
-                use ty::*;
-                let mut markers = self.shallow_type_markers(key).collect::<FxHashSet<_>>();
-                let dcx = self.tcx().dcx();
-                match key.kind() {
-                    Bool
-                    | Char
-                    | Int(_)
-                    | Uint(_)
-                    | Float(_)
-                    | Foreign(_)
-                    | Str
-                    | FnDef { .. }
-                    | FnPtr { .. }
-                    | Closure { .. }
-                    | Coroutine(..)
-                    | CoroutineWitness(..)
-                    | CoroutineClosure(..)
-                    | Never
-                    | Bound { .. }
-                    | Error(_) => (),
-                    Adt(def, generics) => markers.extend(self.type_markers_for_adt(def, generics)),
-                    Tuple(tys) => {
-                        markers.extend(tys.iter().flat_map(|ty| self.deep_type_markers(ty)))
-                    }
-                    Alias(_, _) => {
-                        dcx.warn(format!("Alias type {key:?} remains. Was not normalized."));
-                        return Box::new([]);
-                    }
-                    Pat(ty, _) => {
-                        return self.deep_type_markers(*ty).into();
-                    }
-                    // We can't track indices so we simply overtaint to the entire array
-                    Array(inner, _) | Slice(inner) => {
-                        markers.extend(self.deep_type_markers(*inner))
-                    }
-                    RawPtr(ty, _) | Ref(_, ty, _) => markers.extend(self.deep_type_markers(*ty)),
-                    Param(_) | Dynamic { .. } => {
-                        dcx.warn(format!("Cannot determine markers for type {key:?}"))
-                    }
-                    Placeholder(_) | Infer(_) => {
-                        dcx.fatal(format!("Did not expect this type here {key:?}"))
-                    }
-                }
-                markers.into_iter().collect()
-            })
-            .map_or(&[], Box::as_ref)
-    }
-
-    fn type_markers_for_adt<'a>(
-        &'a self,
-        adt: &'a ty::AdtDef<'tcx>,
-        generics: &'tcx ty::List<ty::GenericArg<'tcx>>,
-    ) -> impl Iterator<Item = &'a TypeMarkerElem> + use<'tcx, 'a> {
-        if adt.is_phantom_data() {
-            Either::Left(
-                generics
-                    .iter()
-                    .filter_map(|g| g.as_type())
-                    .flat_map(|t| self.deep_type_markers(t)),
-            )
-        } else {
-            let tcx = self.tcx();
-            Either::Right(
-                adt.variants().iter_enumerated().flat_map(move |(_, vdef)| {
-                    vdef.fields.iter_enumerated().flat_map(move |(_, fdef)| {
-                        let f_ty = fdef.ty(tcx, generics);
-                        self.deep_type_markers(f_ty)
-                    })
-                }), //.collect::<Vec<_>>(), //.into_iter(),
-            )
-        }
-        .into_iter()
-    }
-
-    pub fn type_has_surface_markers(&self, ty: ty::Ty) -> Option<DefId> {
-        let def_id = ty.defid()?;
-        self.combined_markers(def_id)
-            .next()
-            .is_some()
-            .then_some(def_id)
     }
 
     /// All markers placed on this function, directly or through the type plus
